moving events rsvps into a better file structure; adding files for events rsvps tests
