+++ hknweb/events/tests/views/rsvp_transactions/test_rsvp.py
@@ -1,0 +1,10 @@
+from django.test import TestCase
+
+from django.urls import reverse
+
+from hknweb.events.tests.models.utils import ModelFactory
+
+
+class RsvpViewTests(TestCase):
+    def test_no_event_types(self):
+        pass
