+++ hknweb/events/views/__init__.py
@@ -1,11 +1,13 @@
 from hknweb.events.views.index import index
+from hknweb.events.views.rsvp_transactions import (
+    rsvp,
+    unrsvp,
+    confirm_rsvp,
+)
 
 from hknweb.events.views.views import (
     AllRsvpsView,
     show_details,
-    rsvp,
-    unrsvp,
     add_event,
-    confirm_rsvp,
     EventUpdateView,
 )
